+++ deploy/scripts/connect-db.sh
@@ -1,0 +1,3 @@
+
+#!/bin/bash
+kubectl exec -it -n cafe-boreal deployment/postgres -- psql -U postgres -d cafeboreal